// app/generated/petstore.schemas.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Swagger Petstore
 * OpenAPI spec version: 1.0.0
 */
export interface Pet {
  id: number;
  name: string;
  tag?: string;
}

export type Pets = Pet[];

export type ListPetsParams = {
/**
 * 取得する最大件数
 */
limit?: number;
};